+++ src/main/java/com/buuz135/industrial/proxy/BlockRegistry.java
@@ -70,6 +70,10 @@ public class BlockRegistry {
     public static FluidPumpBlock fluidPumpBlock = new FluidPumpBlock();
     public static FluidCrafterBlock fluidCrafterBlock = new FluidCrafterBlock();
     public static PlantInteractorBlock plantInteractorBlock = new PlantInteractorBlock();
+    public static ItemSplitterBlock itemSplitterBlock = new ItemSplitterBlock();
+    public static FluidDictionaryConverterBlock fluidDictionaryConverterBlock = new FluidDictionaryConverterBlock();
+
+
     public static IFCustomFluidBlock BLOCK_ESSENCE = new IFCustomFluidBlock(FluidsRegistry.ESSENCE, Material.WATER);
     public static IFCustomFluidBlock BLOCK_MILK = new IFCustomFluidBlock(FluidsRegistry.MILK, Material.WATER);
     public static IFCustomFluidBlock BLOCK_MEAT = new IFCustomFluidBlock(FluidsRegistry.MEAT, Material.WATER);
@@ -89,7 +93,7 @@ public class BlockRegistry {
         }
     }.setTickRandomly(true);
     public static IFCustomFluidBlock BLOCK_PROTEIN = new IFCustomFluidBlock(FluidsRegistry.PROTEIN, Material.WATER);
-    private static ItemSplitterBlock itemSplitterBlock = new ItemSplitterBlock();
+
 
     public static void createRecipes() {
         CustomOrientedBlock.blockList.forEach(CustomOrientedBlock::createRecipe);