+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 + Horses and llamas can be tamed again but they need to be tamed first as MC rules
 + Fixed Villager Trade Exchanger doing stuff when it doesn't need to do stuff and fixed missing button translations
 + Added recipe integrations for EnderIO (if present) all the recipes integrations can be disabled in the config
++ Added Fluid Dictionary Converter that will transform fluids into another similar fluids with Crafttweaker integration
+```
+import mods.industrialforegoing.FluidDictionary;
+
+FluidDictionary.add("essence","xpjuice",1);
+FluidDictionary.remove("essence","xpjuice");
+```
 
 # 1.8.7
 + Fixed enchantment extractor leaving empty enchantments tags